# lib/api-zod/src/generated/types/connectPlatformRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ConnectPlatformRequest {
  username: string;
  followerCount?: number;
  avatarUrl?: string;
}