# lib/api-zod/src/generated/types/optimalTimesRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AutoFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OptimalTimesRequestPlatform } from "./optimalTimesRequestPlatform";

export interface OptimalTimesRequest {
  platform: OptimalTimesRequestPlatform;
  category?: string;
  timezone?: string;
}